@@ -1,28 +1,41 @@
 import React from 'react'
 import {Provider} from 'react-redux'
 import {BrowserRouter, Route, Switch} from 'react-router-dom'
+import {ThemeProvider} from 'styled-components'
 
-import DataTable from './components/common/DataTable.container'
-import ResourcesList from './components/resources/ResourcesList.container'
-import ResourcesDetail from './components/resources/ResourcesDetail.container'
-import ResourcesNew from './components/resources/ResourcesNew.container'
-import AuthenticatedApp from './components/App'
+import Home from './App'
+import {withLayout, NotFound} from './components/layout'
+import {
+    AuthProvider,
+    LoginContainer,
+    PasswordHelpContainer,
+    RegistrationContainer,
+    RegistrationConfirmationContainer,
+    RegisterApplicationContainer
+} from './components/auth'
+
+import {withLoginEnhancers} from './components/auth/enhancers'
+import {withAuthentication} from './components/auth/decorators'
 
 import store from './store'
+import theme from './theme'
 
 export default (
-    <Provider store={store}>
-        <AuthenticatedApp>
-            <BrowserRouter>
-                <Switch>
-                    <Route exact path="/" component={ResourcesList} />
-                    <Route exact path="/home" component={ResourcesList} />
-                    <Route exact path="/resources" component={ResourcesList} />
-                    <Route exact path="/resources/new" component={ResourcesNew} />
-                    <Route exact path="/resources/:id" component={ResourcesDetail} />
-                    <Route exact path="/data_table" component={DataTable} />
-                </Switch>
-            </BrowserRouter>
-        </AuthenticatedApp>
-    </Provider>
+    <ThemeProvider theme={theme}>
+        <Provider store={store}>
+            <AuthProvider>
+                <BrowserRouter>
+                    <Switch>
+                        <Route exact path="/" component={withAuthentication(withLayout(Home))} />
+                        <Route exact path="/login" component={withLoginEnhancers(LoginContainer)} />
+                        <Route exact path="/password-help" component={PasswordHelpContainer} />
+                        <Route exact path="/register" component={RegistrationContainer} />
+                        <Route exact path="/confirm-registration" component={RegistrationConfirmationContainer} />
+                        <Route exact path="/register-application" component={RegisterApplicationContainer} />
+                        <Route component={NotFound} />
+                    </Switch>
+                </BrowserRouter>
+            </AuthProvider>
+        </Provider>
+    </ThemeProvider>
 )
