@@ -1,29 +1,14 @@
 import {connect} from 'react-redux'
 import {reduxForm} from 'redux-form'
 import {graphql} from 'react-apollo'
-import Validator from 'validatorjs'
 
 import RegisterApplication from './RegisterApplication'
-import {registerApp, cancel} from './actions'
-import constants from './constants'
+import validators from './validators'
 import {REGISTER_APP} from './mutations'
+import ducks from './ducks'
 
-const {applicationRegistration: {rules, messages}} = constants
-
-const validate = values => {
-    const validator = new Validator(values, rules, messages)
-    validator.passes()
-    return validator.errors.all()
-}
-
-const mapDispatchToProps = dispatch => ({
-    cancel() {
-        return dispatch(cancel())
-    },
-    registerApplication(app) {
-        return dispatch(registerApp(app))
-    }
-})
+const {applicationRegistration: {validate}} = validators
+const {creators: {registerApp}} = ducks
 
 const FormedApplication = reduxForm({
     validate,
@@ -35,9 +20,9 @@ const RegisterApplicationWithData = graphql(REGISTER_APP, {
     props: ({mutate, ownProps}) => ({
         async tryRegisterApp(app) {
             const success = await mutate({variables: app})
-            if (success) ownProps.registerApplication(app)
+            if (success) ownProps.registerApp(app)
         }
     })
 })(FormedApplication)
 
-export default connect(null, mapDispatchToProps)(RegisterApplicationWithData)
+export default connect(null, {registerApp})(RegisterApplicationWithData)
