@@ -1,36 +1,18 @@
 import {connect} from 'react-redux'
 import {reduxForm} from 'redux-form'
 import {graphql} from 'react-apollo'
-import Validator from 'validatorjs'
 
 import Registration from './Registration'
-import {handleError, cancel, register} from './actions'
-import constants from './constants'
+import validators from './validators'
 import {REGISTER_USER} from './mutations'
+import ducks from './ducks'
 
-const {userRegistration: {rules, messages}} = constants
-
-const validate = values => {
-    const validator = new Validator(values, rules, messages)
-    validator.passes()
-    return validator.errors.all()
-}
-
-const mapDispatchToProps = dispatch => ({
-    handleError(error) {
-        return dispatch(handleError(error))
-    },
-    cancel() {
-        return dispatch(cancel())
-    },
-    registerUser(user) {
-        return dispatch(register(user))
-    }
-})
+const {userRegistration: {validate}} = validators
+const {creators: {handleError, register}} = ducks
 
 const FormedRegistration = reduxForm({
     validate,
-    fields: ['password', 'email', 'name', 'grantType'],
+    fields: ['email', 'name'],
     form: 'RegistrationForm'
 })(Registration)
 
@@ -42,7 +24,7 @@ const RegistrationWithData = graphql(REGISTER_USER, {
                 if (error) {
                     throw new Error(error)
                 }
-                if (registerUser) ownProps.registerUser(user)
+                if (registerUser) ownProps.register(user)
             } catch (err) {
                 ownProps.handleError(err)
             }
@@ -50,4 +32,4 @@ const RegistrationWithData = graphql(REGISTER_USER, {
     })
 })(FormedRegistration)
 
-export default connect(null, mapDispatchToProps)(RegistrationWithData)
+export default connect(null, {handleError, register})(RegistrationWithData)
