@@ -2,15 +2,13 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import Button from '../common/Button'
-import Form from '../common/Form'
-import SimpleSvgIcon from '../common/SimpleSvgIcon'
-import ReduxFormField from '../common/ReduxFormField'
-import {breakpoints, forms} from '../common/constants'
+import Link from 'react-router-dom/Link'
+import {Button, LinkButton, Form, SimpleSvgIcon, ReduxFormField} from '../common'
+import {getThemeProp} from '../common/helpers'
 
 const StyledForm = styled(Form)`
     & > * {
-        margin: ${forms.formItemMargin};
+        margin: ${getThemeProp(['forms', 'formItemMargin'], '5px')};
     }
 
     & .attainiaLogo {
@@ -26,7 +24,6 @@ const StyledForm = styled(Form)`
         .attainiaLogo,
         .instructions,
         .name,
-        .password,
         .email,
         .cancelButton {
             max-width: 50em;
@@ -35,11 +32,10 @@ const StyledForm = styled(Form)`
     }
 
     @supports (display: grid) {
-        @media ${breakpoints.desktop} {
+        @media ${getThemeProp(['breakpoints', 'tablet'], 'screen and (min-width: 768px)')} {
             display: grid;
             grid-template-columns: 1fr 1fr;
-            grid-template-areas: 'header header' 'instructions instructions' 'name name' 'email email'
-                'password password' 'save cancel';
+            grid-template-areas: 'header header' 'instructions instructions' 'name name' 'email email' 'save cancel';
 
             & .attainiaLogo {
                 grid-area: header;
@@ -51,10 +47,6 @@ const StyledForm = styled(Form)`
 
             & .name {
                 grid-area: name;
-            }
-
-            & .password {
-                grid-area: password;
             }
 
             & .email {
@@ -71,25 +63,37 @@ const StyledForm = styled(Form)`
         }
     }
 `
-const Registration = ({handleSubmit, tryRegister, cancel}) =>
+const Registration = ({handleSubmit, tryRegister, formCaption, registerLabel}) =>
     <StyledForm onSubmit={handleSubmit(tryRegister)}>
-        <SimpleSvgIcon className="attainiaLogo" width="161" height="39" icon="attainia_logo" />
-        <p className="instructions">Register Your Account</p>
-        <ReduxFormField className="name" placeholder="name" name="name" />
-        <ReduxFormField className="email" placeholder="email" name="email" type="email" />
-        <ReduxFormField className="password" placeholder="password" type="password" name="password" />
-        <Button className="registrationButton" type="submit">
-            Register
-        </Button>
-        <Button className="cancelButton" type="button" onClick={cancel}>
-            Cancel
-        </Button>
+        <SimpleSvgIcon className="attainiaLogo" width="161" height="39" icon="primary" />
+        <p className="instructions">{formCaption}</p>
+        <ReduxFormField
+          id="RegistrationForm-name"
+          className="name"
+          placeholder="name"
+          name="name"
+        />
+        <ReduxFormField
+          id="RegistrationForm-email"
+          className="email"
+          placeholder="email"
+          name="email"
+          type="email"
+        />
+        <Button className="registrationButton" type="submit">{registerLabel}</Button>
+        <LinkButton className="cancelButton"><Link to="/">Cancel</Link></LinkButton>
     </StyledForm>
 
 Registration.propTypes = {
     handleSubmit: PropTypes.func.isRequired,
-    cancel: PropTypes.func.isRequired,
-    tryRegister: PropTypes.func.isRequired
+    tryRegister: PropTypes.func.isRequired,
+    registerLabel: PropTypes.string.isRequired,
+    formCaption: PropTypes.string
+}
+
+Registration.defaultProps = {
+    registerLabel: 'Register',
+    formCaption: 'New User Registration'
 }
 
 export default Registration
