@@ -1,24 +1,58 @@
-import {path} from 'ramda'
+import React from 'react'
+import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import {compose, graphql, withApollo} from 'react-apollo'
+import {compose, withApollo} from 'react-apollo'
 
-import {handleError, logout, login} from './actions'
+import {withStatics} from '../common/helpers'
 import {VALIDATE_TOKEN} from './queries'
+import {removeToken} from './helpers'
 import Validator from './Validator'
+import ducks from './ducks'
 
-const mapStateToProps = state => ({
-    noUserId: !path(['auth', 'user', 'id'], state)
+const {selectors, creators: {logout, handleError, validatedToken}} = ducks
+const mapStateToProps = state => ({token: selectors.parsedToken(state)})
+const mapDispatchToProps = {logout, handleError, validatedToken}
+const mergeProps = (stateProps, dispatchProps, ownProps) => ({
+    ...ownProps,
+    ...stateProps,
+    async tryValidateToken(token) {
+        try {
+            const {error, data} = await ownProps.client.query({
+                query: VALIDATE_TOKEN,
+                variables: {token}
+            })
+            if (error) {
+                throw new Error(error)
+            }
+            if (!data.validateToken) {
+                dispatchProps.logout(token)
+                removeToken(token)
+            } else {
+                dispatchProps.validatedToken(token)
+            }
+        } catch (e) {
+            dispatchProps.handleError(e)
+        }
+    }
 })
 
-export default compose(
-    withApollo,
-    graphql(VALIDATE_TOKEN, {
-        options: ({token}) => ({
-            variables: {token}
-        })
-    }),
-    connect(
-        mapStateToProps,
-        {logout, login, handleError}
+const withReduxConnect = () => connect(mapStateToProps, mapDispatchToProps, mergeProps)
+
+export const withTokenValidation = (DecoratedComponent) => {
+    const WithTokenValidation = ({token, tryValidateToken, ...passThroughProps}) =>
+        <Validator {...{token, tryValidateToken}}>
+            <DecoratedComponent {...passThroughProps} />
+        </Validator>
+
+    WithTokenValidation.propTypes = {
+        tryValidateToken: PropTypes.func,
+        token: PropTypes.string
+    }
+
+    return withStatics(
+        compose(withApollo, withReduxConnect())(WithTokenValidation),
+        DecoratedComponent
     )
-)(Validator)
+}
+
+export default compose(withApollo, withReduxConnect())(Validator)
