@@ -1,59 +1,83 @@
 import uuid from 'uuid/v4'
 import React from 'react'
-import styled from 'styled-components'
+import styled, {withTheme} from 'styled-components'
 import PropTypes from 'prop-types'
-import {colors} from '../common/constants'
+import NavLink from 'react-router-dom/NavLink'
+import {SimpleSvgIcon} from '../common'
+import {getThemeProp} from '../common/helpers'
 
 const Li = styled.li` 
     transition: background 0.1s ease;
     cursor: pointer;
-    color: white;
+    color: ${getThemeProp(['colors', 'grayscale', 'white'], 'white')};
     list-style: none;
     font-size: 16px;
     line-height: 19px;
-    padding-top: 10px;
-    padding-bottom: 7px;
-    background: ${props => props.isSelected && colors.cornflowerBlue};
     border-left-width: 5px;
     border-color: transparent;
     border-left-style: solid;
 
     &:hover {
-        border-color: ${colors.lapisLazuli};
-        background: ${colors.jet};
+        border-color: ${getThemeProp(['colors', 'secondary', 'default'], 'royalblue')};
+        background: ${getThemeProp(['colors', 'grayscale', 'dk'], 'darkgray')};
     }
     & a {
         padding: 10px 15px;
-        color: white;
+        color: ${getThemeProp(['colors', 'grayscale', 'white'], 'white')};
         text-decoration: none;
+        @supports not (display: grid) {
+            display: block;
+        }
+        @supports (display: grid) {
+            display: grid;
+            grid-column-gap: 8px;
+            grid-template-areas: "icon text";
+        }
+        justify-content: start;
+    }
+
+    & a.active {
+        background: ${getThemeProp(['colors', 'secondary', 'default'], 'royalblue')};
     }
 `
 const Ul = styled.ul`
-    background-color: ${colors.outerSpace};
+    background-color: ${getThemeProp(['colors', 'grayscale', 'dk'], 'darkgray')};
     padding: 0;
     margin: 0;
     width: 200px;
     box-sizing: border-box;
 `
 
-const NavBarList = ({items}) => (
+const NavBarList = ({items, theme}) => (
     <Ul>
-        {items.map(({imgSrc, uri, label}) =>
+        {items.map(({iconName, link, label}) =>
             <Li key={uuid()} role="presentation">
-                <a href={uri}>
-                    {imgSrc && <img alt="left" src={imgSrc} />}
+                <NavLink to={link}>
+                    {iconName &&
+                        <SimpleSvgIcon
+                          icon={iconName}
+                          fill={getThemeProp(['colors', 'grayscale', 'white'], 'white')({theme})}
+                        />
+                    }
                     <span>{label}</span>
-                </a>
+                </NavLink>
             </Li>
         )}
     </Ul>
 )
 
 NavBarList.propTypes = {
+    theme: PropTypes.shape({
+        colors: PropTypes.shape({
+            grayscale: PropTypes.shape({
+                white: PropTypes.string
+            })
+        })
+    }),
     items: PropTypes.arrayOf(PropTypes.shape({
-        label: PropTypes.string,
-        imgSrc: PropTypes.string,
-        uri: PropTypes.string
+        iconName: PropTypes.string,
+        label: PropTypes.string.isRequired,
+        link: PropTypes.string.isRequired
     }))
 }
 
@@ -61,4 +85,4 @@ NavBarList.defaultProps = {
     items: []
 }
 
-export default NavBarList
+export default withTheme(NavBarList)
