@@ -2,4 +2,16 @@ import {render} from 'react-dom'
 import Routes from './Routes'
 import './globals'
 
-render(Routes, document.querySelector('#root'))
+function renderApp(Component) {
+    render(Component, document.querySelector('#root'))
+}
+
+renderApp(Routes)
+
+if (module && module.hot) {
+    module.hot.accept('./Routes', () => {
+        /* eslint "global-require": "off" */
+        const DevRoutes = require('./Routes').default
+        renderApp(DevRoutes)
+    })
+}
