@@ -2,12 +2,10 @@ import {combineReducers} from 'redux'
 import {routerReducer as routing} from 'react-router-redux'
 import {reducer as form} from 'redux-form'
 
-import auth from './components/auth/reducer'
-import resources from './components/resources/reducer'
+import authDucks from './components/auth/ducks'
 
 export default combineReducers({
     form,
-    auth,
-    resources,
+    auth: authDucks.reducer,
     routing
 })
